projects/online-anomaly-demo/pyproject.toml
[project]
name = "online-anomaly-demo"
version = "0.1.0"
description = "Multimodal streaming OOD anomaly detection PoC"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "dagster>=1.12.21",
    "datasets>=4.8.4",
    "faiss-cpu>=1.8.0",
    "gradio>=6.10.0",
    "hdbscan>=0.8.40",
    "langgraph>=1.1.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "pyarrow>=20.0.0",
    "scikit-learn>=1.7.2",
    "tqdm>=4.67.1",
    "transformers>=4.57.0",
    "umap-learn>=0.5.9.post2",
    "plotly>=6.6.0",
    "timm>=1.0.26",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
]

[tool.uv]
package = true
